Interface Documentation

This page describes the finite horizon interface. In order to fully define one the user has to also use methods from POMDPs.jl interface defined here.

Docstrings for FiniteHorizonPOMDPs.jl interface members can be accessed through Julia's built-in documentation system or in the list below.

Contents

Interface Documentation

Contents

Index

Types

Model Functions

Index

FiniteHorizonPOMDPs.FiniteHorizon

FiniteHorizonPOMDPs.HorizonLength

FiniteHorizonPOMDPs.InfiniteHorizon

FiniteHorizonPOMDPs.fixhorizon

FiniteHorizonPOMDPs.horizon

FiniteHorizonPOMDPs.stage

FiniteHorizonPOMDPs.stage_stateindex

FiniteHorizonPOMDPs.stage_states

Types

FiniteHorizonPOMDPs.FiniteHorizon — Type

If HorizonLength(m::Union{MDP,POMDP}) == FiniteHorizon(), horizon(m) should be implemented and return an integer

FiniteHorizonPOMDPs.HorizonLength — Type

HorizonLength(::Type{<:Union{MDP,POMDP})
HorizonLength(::Union{MDP,POMDP})

Check whether MDP is Finite or Infinite Horizon and return corresponding struct (FiniteHorizon or InfiniteHorizon).

FiniteHorizonPOMDPs.InfiniteHorizon — Type

HorizonLength(m::Union{MDP,POMDP}) == InfiniteHorizon() indicates that horizon(m) should not be called.

Model Functions

FiniteHorizonPOMDPs.horizon — Function

Return the number of steps that will be taken in the (PO)MDP, given it is Finite Horizon.

A simulation of a (PO)MDP with horizon(m) == d should contain d+1 states and d actions and rewards.

FiniteHorizonPOMDPs.fixhorizon — Function

fixhorizon(m::Union{MDP,POMDP}, horizon::Int)

Wrap infinite horizon (PO)MDP m and horizon to the new structure creating Finite Horizon (PO)MDP.

FiniteHorizonPOMDPs.stage — Function

stage(m::Union{MDP,POMDP}, ss::MDPState)::Int

Return number of state's stage

FiniteHorizonPOMDPs.stage_stateindex — Function

stage_stateindex(m::Union{MDP,POMDP}, ss::MDPState}::Int

Compute the index of the given state in Infinite Horizon for given stage state space.

FiniteHorizonPOMDPs.stage_states — Function

stage_states(m::Union{MDP,POMDP}, stage::Int)

Create Infinite Horizon MDP's states for given stage.